+++ docs/conf.py
@@ -53,6 +53,7 @@ manpages_url = 'https://docs.sysrq.in/{path}'
 
 extlinks = {
     'bug': ('https://bugs.sysrq.in/show_bug.cgi?id=%s', 'bug #%s'),
+    'gentoobug': ('https://bugs.gentoo.org/%s', 'Gentoo bug #%s'),
     'gitweb': (f'https://git.sysrq.in/{project}/tree/%s', '%s'),
     'pypi': ('https://pypi.org/project/%s/', '%s'),
 }